refactor: Move data type declarations to own file
Move Post, Comment, and Category types to a types file since they are
needed in more than one place and have to avoid circular dependencies.
